switch to unbounded channels & notify
it seems bounded channels caused some backpressure bug when training
very fast.
rather than trying to pick a perfect channel size, we just let it grow.
if your cpu can't keep up with your training, an OOM is fine.

additionally, we replace a channel of () with a tokio::sync::Notify for
simplicity.