implement allowlist into clients
rather than joining all p2p peers the server knows,
we now pull our p2p peers from the coordinator's current state.
if we have a new peer in the coordinator's clients list,
we connect at that moment in time.

*NOTE:* this does not force-disconnect peers who are no longer in
the allowlist. TODO for later.

in the future, this refactor would clean the code up:
https://github.com/NousResearch/psyche/issues/119