Update coordinator state every waiting for members tick (#281)
* Update coordinator state for every waiting for members tick
If all clients leave mid run and another one joins then it will contain and stale coordinator stale and it will not be able to join the run since the get_active_clients function will not detect this new client. To fix it we just have to update the coordinator state everytime in waiting for members to get the latest state.

* Fix flag restart to join new epochs
Also the flag to detect if the client already joined the next run or not was not being reset correctly, we included as part of the client and reset it if was previously set