fix: witness could get sent early + int underflow
when in a step, the .clients list can lose a client.
if that happens, we could send the witness early, before some client has
had the chance to send finished.

we could also int underflow.

to fix this, we actually check each active client's finished messages
instead of just comparing length of the sets.