p2p: allow fallback nodes in download
previously, if a node went offline after their blob was downloaded by
some witnesses and not others, we could get a consensus failure.
now, we simply pass every single node as a potential download source;
so if the data makes it to one single witness, it *should* get
propagated across the entire network.
if nobody has it, this should fail as per usual.